Move osu api interfaces to separate file
